// server/src/env/types.ts
export enum Env {
  NODE_ENV = 'NODE_ENV',
  COOKIE_SECRET = 'COOKIE_SECRET',
  CSRF_TOKEN_SECRET = 'CSRF_TOKEN_SECRET',
  MONGO_URI = 'MONGO_URI',
  REDIS_HOST = 'REDIS_HOST',
  REDIS_PORT = 'REDIS_PORT',
  REDIS_CACHE_TTL = 'REDIS_CACHE_TTL',
  PASSWORD_SALT_ROUNDS = 'PASSWORD_SALT_ROUNDS',
  GOOGLE_ADMIN_EMAIL_ADDRESS = 'GOOGLE_ADMIN_EMAIL_ADDRESS',
  GOOGLE_CLIENT_ID = 'GOOGLE_CLIENT_ID',
  GOOGLE_CLIENT_SECRET = 'GOOGLE_CLIENT_SECRET',
  GOOGLE_MAILER_REFRESH_TOKEN = 'GOOGLE_MAILER_REFRESH_TOKEN',
  OTP_SALT = 'OTP_SALT',
  TOKEN_SECRET = 'TOKEN_SECRET',
}